Don't bind to web port until application is deployed
This fixes an issue where the first request of a new dyno always fails
on Heroku because JBossWeb was binding to the web port before the web
context was actually registered, resulting in a 404 for the initial
request.

This temporarily requires depending on a TorqueBox 2.3.0-SNAPSHOT
binary distribution, so if you're building torquebox-lite from source
you'll first need to build TorqueBox from source. As soon as TorqueBox
2.3.0 comes out it will be updated to remove the -SNAPSHOT.